Backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dacite==1.9.2
Deprecated==1.2.18
fastapi==0.103.2
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.3.0
gensim==4.3.3
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lap==0.5.12
lyricsgenius==3.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
mutagen==1.47.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
psutil==7.0.0
py-cpuinfo==9.0.0
pycryptodomex==3.22.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
pykakasi==2.3.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytube==15.0.0
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
setuptools==77.0.3
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundcloud-v2==1.6.0
soupsieve==2.6
spotdl==4.2.11
spotipy==2.25.0
starlette==0.27.0
sympy==1.13.1
syncedlyrics==1.0.1
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.95
ultralytics-thop==2.0.14
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.23.2
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
yt-dlp==2025.3.31
ytmusicapi==1.10.1